Extract `AnimationFileData` structs to separate header file
Use forward declaration for `AnimationFileData` in `AnimationFile` header.

We still provide public access to the parsed data `struct` definitions. The `AnimationFile` accessor methods can still provide direct access to the parsed data. This may be useful in the future should we ever want to provide support to edit and write back updated data to a new file.
