+++ NAS2D/NAS2D.vcxproj
@@ -245,6 +245,7 @@
     <ClInclude Include="Renderer\Window.h" />
     <ClInclude Include="Resource\ResourceCache.h" />
     <ClInclude Include="Resource\AnimationFile.h" />
+    <ClInclude Include="Resource\AnimationFileData.h" />
     <ClInclude Include="Resource\AnimationFrame.h" />
     <ClInclude Include="Resource\AnimationSequence.h" />
     <ClInclude Include="Resource\AnimationSet.h" />
